refactor: combine imageutils and videoutils into a single mediautils
